// client/src/pages/admins/helpers/_schemas.tsx
export const EditAddressSchema = yup.object().shape({
    address_1: yup.string().min(3, "Longitud minima 3 digitos").required("Campo obligatorio"),
    address_2: yup.string().optional(),
    city: yup.string().min(3, "Longitud minima 3 digitos").required("Campo obligatorio"),
    state: yup.string().required("Campo obligatorio").notOneOf([""]),
    zip: yup.string().min(3, "Longitud minima 3 digitos").required("Campo obligatorio"),
})

export const initialValuesAddress = {
    address_1: "",
    address_2: "",
    city: "",
    state: "",
    zip: "",
    tipo: ""
}